% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcht.R
\name{searcht}
\alias{searcht}
\title{search the cut-off point controlling FDR}
\usage{
searcht(l1, eta, p)
}
\arguments{
\item{l1}{A list of test statistics.}

\item{eta}{Desired FDR level.}

\item{p}{Variable dimension.}
}
\description{
This function allows you to find the t_hat
}
