% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatlist.R
\name{teststatlist}
\alias{teststatlist}
\title{get the fdr and power for the two-stage method}
\usage{
teststatlist(n, p, eta, maxbeta, alpha)
}
\arguments{
\item{n}{Sample size}

\item{p}{Variable dimension.}

\item{eta}{Desired FDR level.}

\item{maxbeta}{Signal size}

\item{alpha}{Tuning parameter for stage 1}
}
\description{
This function allows you to get the fdr and power for the two-stage method
}
